+++ wire-ios-request-strategy/Sources/Message Sending/MessageInfoExtractor.swift
@@ -18,59 +18,6 @@
 
 import Foundation
 
-struct UserClientData: Equatable {
-    var sessionID: ProteusSessionID
-    var data: Data?
-}
-
-/// Container of information needed to encrypt a message
-struct MessageInfo {
-    typealias Domain = String
-    typealias UserID = UUID
-    typealias ClientList = [Domain: [UserID: [UserClientData]]]
-
-    var genericMessage: GenericMessage
-    /// list of clients divided per domain and userId
-    var listClients: ClientList
-    var missingClientsStrategy: MissingClientsStrategy
-    var selfClientID: String
-    var nativePush: Bool
-    private var userClients: [UserClient]
-    
-    internal init(genericMessage: GenericMessage,
-                  listClients: MessageInfo.ClientList,
-                  missingClientsStrategy: MissingClientsStrategy,
-                  selfClientID: String,
-                  nativePush: Bool,
-                  userClients: [UserClient]
-    ) {
-        self.genericMessage = genericMessage
-        self.listClients = listClients
-        self.missingClientsStrategy = missingClientsStrategy
-        self.selfClientID = selfClientID
-        self.nativePush = nativePush
-        self.userClients = userClients
-    }
-    
-    func allSessionIds() -> [ProteusSessionID] {
-        var result = [ProteusSessionID]()
-        for (_, userClientIdAndSessionIds) in listClients {
-            for (_, userClientDatas) in userClientIdAndSessionIds {
-                let sessionIds = userClientDatas.compactMap( { $0.data == nil ? $0.sessionID : nil })
-                result.append(contentsOf: sessionIds)
-            }
-        }
-        return result
-    }
-    
-    func resetAllUserClientsFailedSessions(in context: NSManagedObjectContext) async {
-        await context.perform {
-            userClients.forEach { $0.failedToEstablishSession = false }
-            context.saveOrRollback()
-        }
-    }
-}
-
 enum MessageInfoExtractorError: Error {
     case missingConversation
     case missingGenericMessage
